updated landing page
