+++ apps/betterangels/package.json
@@ -85,7 +85,8 @@
     "@swc/helpers": "*",
     "graphql": "*",
     "vite": "*",
-    "@nx/vite": "*"
+    "@nx/vite": "*",
+    "vitest": "*"
   },
   "scripts": {
     "eas-build-pre-install": "corepack enable && mkdir -p .yarn && cp -R ../../.yarn/patches .yarn/patches 2>/dev/null || :",
