--- apps/betterangels/src/app/(tabs)/clients.tsx
@@ -1,13 +1,0 @@
-import { Clients, ClientsAddInteraction } from '@monorepo/expo/betterangels';
-import { useLocalSearchParams } from 'expo-router';
-import Logo from '../assets/images/logo.svg';
-
-export default function ClientsScreen() {
-  const { createInteraction } = useLocalSearchParams();
-
-  if (createInteraction) {
-    return <ClientsAddInteraction Logo={Logo} />;
-  }
-
-  return <Clients Logo={Logo} />;
-}
