+++ apps/betterangels/src/app/(tabs)/interactions.tsx
@@ -1,5 +1,6 @@
 import { Interactions } from '@monorepo/expo/betterangels';
+import Logo from '../assets/images/logo.svg';
 
 export default function InteractionsScreen() {
-  return <Interactions />;
+  return <Interactions Logo={Logo} />;
 }
