--- libs/expo/betterangels/src/lib/screens/Home/ActiveClients.graphql
@@ -1,35 +1,0 @@
-query ClientProfiles(
-  $filters: ClientProfileFilter
-  $pagination: OffsetPaginationInput
-  $order: ClientProfileOrder
-) {
-  clientProfiles(filters: $filters, pagination: $pagination, order: $order) {
-    totalCount
-    pageInfo {
-      limit
-      offset
-    }
-    results {
-      id
-      age
-      dateOfBirth
-      email
-      firstName
-      heightInInches
-      lastName
-      mailingAddress
-      middleName
-      nickname
-      residenceAddress
-      hmisProfiles {
-        id
-        agency
-        hmisId
-      }
-      profilePhoto {
-        name
-        url
-      }
-    }
-  }
-}
