+++ libs/expo/betterangels/src/lib/screens/index.ts
@@ -5,7 +5,6 @@ export { default as Clients } from './Clients';
 export * from './ClientsAddInteraction';
 export { default as CreateClientProfile } from './CreateClientProfile';
 export { default as FileScreenComponent } from './FileScreenComponent';
-export { default as Home } from './Home';
 export { default as Interactions } from './Interactions';
 export { default as Note } from './Note';
 export { default as PublicNote } from './PublicNote';
