+++ libs/expo/betterangels/src/lib/ui-components/NavModal.tsx
@@ -1,11 +1,9 @@
 import {
   BarsIcon,
   HouseLineIcon,
-  NoteIcon,
   SettingsOutlineIcon,
   SignOutIcon,
   TaskListIcon,
-  UsersLineIcon,
 } from '@monorepo/expo/shared/icons';
 import { Colors, Radiuses, Spacings } from '@monorepo/expo/shared/static';
 import { Avatar, TextRegular } from '@monorepo/expo/shared/ui-components';
@@ -37,8 +35,6 @@ export default function NavModal(props: INavModalProps) {
 
   const ACTIONS = [
     { title: 'Home', Icon: HouseLineIcon, route: '/' },
-    { title: 'Clients', Icon: UsersLineIcon, route: '/clients' },
-    { title: 'Interactions', Icon: NoteIcon, route: '/interactions' },
     tasksFeatureOn && {
       title: 'Tasks',
       Icon: TaskListIcon,
